requirements.txt
# Compatible LangChain ecosystem
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-text-splitters==0.0.2
langchain-groq==0.1.3

# Vector DB and embeddings
chromadb==0.4.24
sentence-transformers==2.5.1

# Web framework
streamlit==1.32.2

# Data processing
pandas==2.2.1
numpy==1.26.4
requests==2.31.0

# Utilities
python-dotenv==1.0.1
openpyxl==3.1.2

